Move status into package
